+++ packages/client/vite.config.ts
@@ -2,6 +2,7 @@ import { defineConfig } from 'vitest/config';
 
 export default defineConfig({
   test: {
+    testTimeout: 15000,
     coverage: {
       ignoreEmptyLines: true,
       provider: 'v8',